wallet: move CreateImportedAccount across all backends
